Refactoring user login with password
